extracted labelstream to it's own class
