+++ src/app/dataview/databar/databar.component.ts
@@ -16,7 +16,8 @@ import { Spinner } from 'spin.js';
 import { largestTriangleThreeBucket } from 'd3fc-sample';
 import { Sensor } from "../sensor";
 import { Colorer} from '../colorer';
-import { Labeller, Label, LabelStream } from './labeller';
+import { Labeller } from './labeller';
+import { Label, LabelStream } from '../labelstream';
 import { Drawer } from './drawer';
 import { ToolMode, ModeTracker } from '../modes/tool-mode';
 import * as d3 from "d3";
