+++ src/app/dataview/dataview.component.ts
@@ -6,7 +6,7 @@ import { saveAs } from 'file-saver/FileSaver';
 import { DataloaderService } from '../data-loader/data-loader.service';
 import { WorkspaceInfo, DataInfo } from '../data-loader/workspace-info';
 import { SettingsService } from '../settings/settings.service'
-import { Label, LabelStream, EventMap } from './databar/labeller';
+import { Label, LabelStream, EventMap } from './labelstream';
 import { ToolMode, ModeTracker } from './modes/tool-mode';
 import { Colorer } from './colorer';
 import { Sensor } from './sensor';
